Update README.md
This project was created to create a grocery counter with a histogram in the end of the program to illustrate the number of items purchased. This project came out very well with little to no errors. As of right now, I would like two make two changes. First, the exit key could be any character key. I would like to make it specific to one letter like ‘q’. Making this improvement will reduce the chances of accidently hitting a character key and exiting the program. Secondly, I would like to display all the grocery items after each input. Right now, the code only displays the total grocery item of the item that was most recently added. These improvements would help the end user better visualize what items are added to the grocery list. The most challenging code to write was creating a vector pair. Creating a vector pair itself was not difficult but, it was difficult because I could had forgotten about the vector pair capabilities. The great thing about this project is that it can be used in other forms of counters or histogram projects with small modifications to the variables. I did my best to make my code very readable by spacing everything out well and naming the variables identical to the grocery items. The code is built  on a simple while loop and if statements.